requirements.txt
inquirer
mysql-connector-python
